Add documentation for Uninitialized Attribute (code 13)
Summary:
This error isn't hard to document in simple cases,
but the most common cause of it in practice is code using
dataclass-like semantics which are an open problem not just for
pyre but for python typing in general.

Quite a few projects hit this issue (e.g. using python classes with
a decorator that extends dataclass to represent data schemas, various
test frameworks) and there's no way to solve it without a new PEP.

So I decided to openly document that this is a known problem and link to
the active discussions of it, I'm not sure if this is kosher but I think
it may be important for users who need to understand that
 (a) there's no current solution, you have to use a fixme / ignore
 (b) the typing community is aware of their pain

Reviewed By: shannonzhu

Differential Revision: D28508666

fbshipit-source-id: 15e2342d000a1432bd34e5366e057821e7a4f532