Infer: Set up test scaffold
Summary:
Expose some of the functions and types from typeInferenceAnalysis so that
we can use them to wire up single-pass tests of the `analyze` function output.

For now the tests are stubs so that we can focus on the plumbing (how to run
infer in tests, how to throw useful errors if someone gets a module name wrong
when writing a new test, etc); the next diff hooks up `typeInferenceData` and fills
out these tests with actual validations of the inferred types.

Reviewed By: grievejia

Differential Revision: D28491243

fbshipit-source-id: 0a4150c52f8c7df73411e89de0b11da7db2d2030