Infer: move `abstract` into `InferenceResult.t`
Summary:
As I started to look at reporting inference results to the client
side, I realized that we should probably make the `DefineAnnotation.t`
type have exactly the information needed to produce stubs from the
python client.

The `abstract` field is not needed for this - it's an implementation
detail on the ocaml side, only used to prevent us from inferring
NoneType as the return type of abstract methods.

So, I moved it to `InferenceResult.t` which will never actually
be exposed to the client.

I'm also removing the `to_yojson` derivation for `InferenceResult.t` because we
won't need it - the data we serialize will be a `GlobalInferenceResult` type
that I intend to define in the next diff rather than this local inference
result, which is an implementation detail of the local <-> global boundary in
the interprocedural framework.

Reviewed By: grievejia

Differential Revision: D28572007

fbshipit-source-id: 029027d9aefda2566f08d03de0accee45746caf5