Add docs for Typed Dict Key Not Found (code 27)
Summary:
I decided to explicitly state that the typed dict interface currently doesn't
support extra keys (i.e. there's no equivalent of javascript + flow's `...`)
because it's actually not that easy to tell from existing docs whether this is
possible; I didn't see any explicit discussion either way, but nothing
in the PEP or in the `typing` docs makes any mention of a typed dict that's
open to extension.

Also if I'm mistaken and it actually is possible, someone please correct me!

Reviewed By: pradeep90

Differential Revision: D28603468

fbshipit-source-id: 1767ae02490ee42dfef08c7c7de1b96e4cc8c739