Make the shared memory configurable (server side)
Summary:
This diff makes the shared memory configurable in the server. It allows to set the heap size, dependency table power and hash table power by passing command line arguments to the binary.
The follow-up diff will make these options available in the client, using command line options and `.pyre_configuration` options.
The end goal is to lower down the current default heap size (16GB), and set it higher depending on the repository.

Reviewed By: grievejia

Differential Revision: D28546062

fbshipit-source-id: 9a1f3d0139c209279f4dff3a0218ae330cfe3ea4