Rewire interprocedural functions a bit
Summary:
To support infer reporting, I need access to the global resolution
(which lives inside the type environment) at reporting time, because
at least for now I'll need the ability to join types for globals
and attributes if any inferences clash.

While I was making that change, I figured I'd also make the order
of "context" arguments consistent across all our functions in
`interproceduralResult` and `interproceduralAnalysis`, so now the
order is always:
- scheduler
- static_analysis_configuration
- environment

Having a consistent order makes it easier to quickly skim and see
which args are present in which calls; when they are out of order
you have to read the whole arg list more carefully.

Reviewed By: dkgi

Differential Revision: D28612887

fbshipit-source-id: df4402b257472fba0243a6fae19252b72ce58d22