Infer: hook up global result to reporting and update tests
Summary:
This change takes advantage of the preceding diff, which allows
interprocedural analyses to access the environment when serializing
results, to hook up `GlobalResult.t` to reporting.

At this point my plan is actually to not dump to disk at least in the
short term - this is something we want in pysa but the existing
pyre infer communicates over stdout and I think that's good enough
for now.

Reviewed By: dkgi

Differential Revision: D28614405

fbshipit-source-id: bb83220a16ec6f4ffce4582467fa54eb36652d48