Show final information in reveal_type
Summary:
Tweak the code for `reveal_type` to display a `(final)` suffix
for annotations where the mutability is restricted to be final.

This allows users to see the difference between final and non-final
types, which currently print exactly the same thing in `reveal_type`
even though pyre is aware of differences in type information.

**[RFC] What might we want to change here?**

I think my handling of the printing here is simple and better than
what we have, but it's not perfect:
- I skip printing the final modifier in cases where we infer a
  type because that gets verbose, but in some cases we might
  still want to see the final
- related to ^ is the fact that we generally won't want to show
  `final` in the case of Literal type whether inferred or not

So it might be better to rewire the code so that
- the final and annotation handling get smashed together and
  we suppress the `(final)` whenever the annotation is `Literal`
- whenever we are showing an `(inferred: ...)` section, we put
  the final suffix inside there rather than after the original
  type

This would be more complicated though so I'd rather get opinions
before trying to implement it

I also wonder if the pretty-printing ought to be moved into
`annotation.ml` instead of getting inlined into a switch statement
of `analysisError.ml` - I can do that in a separate diff though.

Reviewed By: dkgi

Differential Revision: D28643437

fbshipit-source-id: 214b128b269afb933dfeb514bcedc8e990892deb