Added unit tests for `Polynomial.replace` function
Summary:
Added some unit tests for `Polynomial.replace` function.

They should cover some important cases in the code. This should be expressed in four cases of replacing in a monomial with a monomial, replacing in a polynomial with a monomial, replacing with a polynomial in a monomial, and replacing in a polynomial with a polynomial.

- Replacing a variable that is not present in the expression
- Replacing a variable by zero
- Replacing a variable by a constant
- Replacing a variable by itself
- Proper multiplicative behavior of coefficients and additive behavior of degree when replacing a variable by another monomial
- Proper "isolating" behavior when replacing in multivariate monomials
- Proper distributive behavior when replacing a polynomial for a n-degree variable
- When replacing in a polynomial, monomial terms that are of the same "rank" after replacement are combined properly

Reviewed By: shannonzhu

Differential Revision: D28705831

fbshipit-source-id: 6a7d4e24210fbb4b2bf4de4f61b6d151f42293e1