Added tests for `local_replace_polynomial`
Summary:
Added some tests for the `local_replace_polynomial` function. As it stands currently, polynomial constants are still being un-normalized (to the form of `Type.Literal`, for instance), so that will need to be fixed when I tackle Task 2.

Also to note is that apparently the replacing done occurs in "waves", that is first we find all occurrences of `x` and replace for them, and then all occurrences of `y`, and so on. This is relevant because if you replace with a polynomial that has variables that were in the original polynomial, you may actually replace for the new variables you introduced as well. Unclear if this is intentional.

These tests should handle:
- Replacing in a polynomial that has on variables
- Replacing in a polynomial with a single variable
- Replacing in a polynomial with multiple variables
- Correct merging of monomials of the same "rank" after replacing
- Replacing in a polynomial with another polynomial

Reviewed By: pradeep90

Differential Revision: D28734020

fbshipit-source-id: 72af560de1cd77efa65d21c82b6afbce704ce6ce