Rename inlined wrapper functions.
Summary:
Problem: We want to flatten the list of decorated functions. However, they are currently named as `inner` (the name of the decorator's wrapper function). This makes their qualifier be `foo.inner`, which is not very readable.

So, name them as `foo.__inlined_with_logging`. That way, it's clear that we're referring to the `with_logging` decorator.

Remaining work:
+ Lift nested wrappers into a flat list of decorators.

Reviewed By: arthaud

Differential Revision: D28805597

fbshipit-source-id: 4ab200c3271ad84d1156c0843d68008826a26dd3