Disable partial subtractions of slots in FlowDetails
Summary:
This patch disables the partial substractions of slots in `FlowDetails`, which is used as an optimization to save memory in the tree domain.
This fixes the bug described in D28497136 (https://github.com/facebook/pyre-check/commit/efabcbbb8c96b88d08b0cc42271efaf3b9aa3b14), and possibly other hidden bugs due to subtracting partial slots.
The only downside is performance, which I suspect is minimal in that case since this subtraction is only reached when we have the same leaf (e.g, `UserInput`) and trace info (e.g, `CallSite(function, location)`), which should be rare.

Reviewed By: fahndrich

Differential Revision: D28497579

fbshipit-source-id: b3322aeb183562479d9090608dbc733a6ce8df0c