Add function for uniquifying names.
Summary:
Need this to avoid name clashes among decorators. For example, if we apply the same decorator twice, then having a flat list of inlined wrappers means that they will clash. (Note that their bodies are different - one will call the previous decorated function, so we can't just ignore the name clash.)

There won't be name clashes among helper functions because we nest them within the wrapper.

The ordering is a bit weird because we pass in decorators as `[decorator1, decorator2, decorator3]` and we want to start numbering duplicates from the end.  Let me know if anyone has strong opinions about numbering from the start.

Reviewed By: dkgi

Differential Revision: D28819697

fbshipit-source-id: 637b5c4983b61509a284aa34c14b469581ff9d0a