Infer: build a minimal infer-v2 python command
Summary:
Create a bare-bones python command infer-v2 that knows how to drive
the ocaml-side type inference.

None of the stub-generation and related logic
is available yet, the goal is just to be able to drive
interprocedural infer on real projects - for example I wanted to be
able to see if the runtime is reasonable.

Reviewed By: shannonzhu

Differential Revision: D27631635

fbshipit-source-id: 484d66a7502d44049a57c89cf113ee944a3cec7b