Make the shared memory configurable (client side)
Summary:
This diff adds command line options and configuration options to control the shared memory, including the heap size, dependency table power and hash table power.

The command line options are:
* `--shared-memory-heap-size`
* `--shared-memory-dependency-table-power`
* `--shared-memory-hash-table-power`
The `.pyre_configuration` options are:
```
"shared_memory": {
  "heap_size": <value>,
  "dependency_table_power": <value>,
  "hash_table_power": <value>
}
```
The end goal is to lower down the current default heap size (16GB), and set it higher depending on the repository.

Reviewed By: grievejia

Differential Revision: D28621286

fbshipit-source-id: 8a5358d1b10ec0267f5fd272c7509e66a79294c7