Track the depth of propagations (i.e, tito)
Summary:
Pysa uses a backward analysis to infer whether a tainted parameter would be propagated to the return value. This diff computes the "depth" of that propagation, defined as the minimal number of hops (i.e, function calls) needed to perform that propagation.

For instance:
```
def foo(x):
  return x # depth 0
def bar(x):
  return foo(x) # depth 1
def baz(x):
  return bar(x) if rand() else foo(x) # depth 1 + min(1, 0) = 1
```

In a follow-up, we will limit the depth of propagations that we infer. This should help us fix performance issues and also reduce some false positives. A downside is that this might potentially hide flows, so use this with caution.

Reviewed By: fahndrich

Differential Revision: D28858640

fbshipit-source-id: fe7e66ccd887d24e85e87bc22c27ddb218a17339