Infer: add index and sort-by-index to define params
Summary:
In order to produce a correct python stub for a define, we'll
need to know the parameter order. Record that information and
add sorting to the `ByName` maps.

All the other maps (which don't *need* to be sorted) are now sorted
by location because that will make reading the results easier,
e.g. debugging, since now
- all annotations from a given file will be grouped together
- they will appear in same order they appear in the code

Reviewed By: dkgi

Differential Revision: D28868546

fbshipit-source-id: 6e234001b01234c30550ba1ed832a129145eaa78