Fix bug in handling of optional paths
Summary: This bug surfaced in D28902458. Basically, we weren't properly handling the fact that some paths in model verification errors can be None. In the v1 command, we tried to handle it, but ended up comparing a `Path` to a `str` which caused the exception in D28902458. In the v2 command, we weren't handling the issue, so we were comparing a `Path` to a `None`. This diff fixes both and adds a test.

Reviewed By: dkgi

Differential Revision: D28904666

fbshipit-source-id: 0ddafcc3978a7236aebe0b6387de2de9ce4f8f72