Infer: add minimal stub logic to client
Summary:
This commit add stub generation to our basic python client: we build up annotations
as a tree of dataclasses, and the annotations can be converted into stub file format.

This is a pretty big commit as-is, so I'm not yet adding the actual write-to-disk
portions of the code. I'll put those in place eventually, but for the moment it's
actually easier to focus on the core logic and a test harness and just print
the stubs to stdout.

There are at least two features I'll still need to add which I didn't fully
realize until working on this code:
- a complete_only flag to disable stubs, when requested, that lack some
  parameter annotations
- the ability to add explicit annotations from libcst when necessary, to
  enable full stubs

Reviewed By: dkgi

Differential Revision: D28883959

fbshipit-source-id: 7a20da145d241bc2ca303e822c8af62c878d6f08