Infer: Allow combining two ModuleAnnotations
Summary:
The `__add__` function combines the annotations in two
`ModuleAnnotations` instances, preferring annotations from
the left when there are collisions.

The current purpose is to be able to combine together
existing annotations with the results of pyre infer (preferring infer
output when both are available) when making full stub files.

Reviewed By: pradeep90

Differential Revision: D28935186

fbshipit-source-id: c960ecd4f7976496a3fd96d13aea4aa2384816f1