Changed `IntExpression` to take in abstract type
Summary:
+ We would like to move towards an abstract type for `IntExpression`s to make sure that all polynomials are only expressed in fully normalized form.
+ Currently, an `IntExpression` is of `'a Polynomial.t`, which is transparent. This allows for all kinds of redundant states, like the monomial list (informally) `[1, [x, 1]]`, which is better represented as a `Variable x`, and `[5, []]`, which is better represented as `Literal (Integer 5)`. This is also a gigantic pain because these errors tend to be "hidden", and the error messages you get for mistaking alternate representations are terrible.
+ We have added an abstract type `'a RecordIntExpression.t`, which is the new input to an `IntExpression`. Everywhere the `IntExpression` constructor was used, we now use `IntExpression.create`, which implicitly constructs proper values of the abstract type for use.
It is a common occurrence to pattern match on an `IntExpression _` to extract out the polynomial, and perform some polynomial-only operation on it, so in addition we have added support for the private type `'a RecordIntExpression.t`, which is now supported at all call-sites. This means that we may deconstruct an `IntExpression` to obtain a `Polynomial.t`, but we cannot construct an `IntExpression` using the constructor itself (preserving normalization).
Currently, no normalization is happening, but this change is to stage for later normalization features. In addition, we have a sum type `'a variant`, which is necessary because `IntExpression` regards the type of `RecordIntExpression.t` as abstract, so we must have `RecordIntExpression.t` implement its own normalizing constructor, with a sum to tag whether it is actually a polynomial, or some other type. This will be expanded on in future diffs.

There's some redundant stuff that will need to be taken away in a future diff. For instance, we should get rid of some useless branches and re-evaluate everything to do with `type_to_int_expression`.

- **Why `module IntExpression`?**
Firstly, we need some function of signature `'a Polynomial.t -> type_t` - this function will take a polynomial and return its normalized form (which may be `IntExpression`, `Literal`, or `Variable`). Necessarily, since `RecordIntExpression` comes *before* `type_t` and we are not using recursive modules, this must be in a different module. `IntExpression` is this module.

- **Why `'a variant`?**
`'a variant` is here because, as stated before, `IntExpression` contains the function `create : 'a Polynomial.t -> type_t`. However, the type of `RecordIntExpression` is *abstract* to the module `IntExpression`, because we want to enforce that the entirety of `type.ml` regards `RecordIntExpression.t` as abstract, so we don't have any representation invariant violations. Thus, `RecordIntExpression` must itself implement its own normalization function, with a way of signalling to `IntExpression` whether the output it produced is normalized as a polynomial, or needs to be changed to a `Literal` or something.

- **Why not just have `RecordIntExpression` have a function that turns a polynomial into a pseudo-normalized polynomial, and then have `IntExpression` pattern match on it to produce the proper value?**
In other words, why not have `RecordIntExpression.normalize : 'a Polynomial.t -> 'a Polynomial.t`, and have `create` do the work of recognizing if its a literal or not.
The problem is that `RecordIntExpression` must produce the abstract type that `IntExpression` (the constructor) takes in, since it comes before the type definition of `type_t`. Then, `IntExpression.create` has the burden of creating the value of type `RecordIntExpression.t` - which it can't, since it's an abstract type.
So `RecordIntExpression.t` *has* to give a way to produce values of type `RecordIntExpression.t`, and because those values are abstract, it must tag them with whether or not they should be normalized.
- **Why the `private` constructor for `RecordIntExpression.t`?**
It is entirely possible to implement `RecordIntExpression.polynomial_of : 'a t -> 'a Polynomial.t`. However, this induces extreme headache in the existing codebase when trying to pattern match on `IntExpression` variants of the `type_t` type.
Without a private constructor, this means that to pattern match on an empty polynomial, you must do `| IntExpression int_expression -> match int_expression of | [] -> ...`, or something similar, as opposed to `| IntExpression (Data []) -> ...`. It's just a lot of syntactic burden. This can be alleviated later by reducing the amount of points where we have to pattern match on `IntExpressions`, by changing `solve_less_or_equal` to do all of its simplification first, rather than recursively inside of the function. Right now, however, this is less bloat.

Reviewed By: grievejia

Differential Revision: D28890690

fbshipit-source-id: ef1d81aec74922d9c74f4e34b998d5e07f0fd63d