Rewire interprocedural to be one-Analysis [1] update public interface
Summary:
**About the stack**

After some discussion, all the pysa stakeholders are agreed that we don't have
any expectation of needing to run multiple concurrent analyses and share data
across them.

As a result, we can significantly simplify the framework by restricting
to one analysis at a time.

Code simplification is an important long-term goal, but our immediate concern
is really functionality: by restricting the framework to run on one analysis
at a time, we can make it possible for us to put callbacks into the analysis
which the "plumbing" code, such as callgraph construction, can use to customize
framework behavior for each analysis; this isn't possible to do cleanly in
a multi-analysis framework because it's not clear which analysis would call
the shots.

This change is needed to unblock full feature parity of interprocedural
infer vs ordinary infer, because we need the ability to exclude some code
from the analysis based on commandline flags that don't line up with
taint.

**About this diff**

It's going to take some time to fully simplify the existing multi-analysis
code. This diff is a minimal initial change to rewire the public functions
from `interproceduralAnalysis.ml` while making it trivial to review the
changes; the interfaces are clean but the code is now a little messy
because to preserve the existing logic I'm often doing a "fold" over just one
analysis.

I'll stack changes on top to both simplify the interproceduralResult types
and get rid of some of the accumulate boilerplate still left over because
we previously had to fold over analyses.

Reviewed By: dkgi

Differential Revision: D28886238

fbshipit-source-id: d82d3c422a71c8ce0a260289e0ab36d67e60d356