Rewire interprocedural to be one-Analysis [2] simplify accumulations
Summary:
In the previous diff, to keep the review easy I made minimal changes to
functions' top-level signatures, but preserved the inner logic which left
a bunch of functions folding over just one abstract kind.

In this diff I clean that up - the resulting functions are identical but
shorter and easier to read.

I'm not yet changing `model_t` and `result_t` to contain just a single
analysis because that will be a big change, and I'd prefer to have things
as clean as possible beforehand to minimize confusion. I'll stack that change
on top which will further simplify the code.

Reviewed By: dkgi

Differential Revision: D28886237

fbshipit-source-id: 722833a8d7d0e7f8e1384dcc41eb873a0d764930