Treat parse errors in pysa files as model verification errors (closes #431)
Summary: We currently throw an exception on syntax errors in pysa files. This means that `pyre query validate_taint_models()` returns an error with a different format than a regular model verification error. This is problematic for the VSCode plugin we are trying to implement, which relies on `pyre query` returning errors in a given format.

Reviewed By: dkgi

Differential Revision: D28950011

fbshipit-source-id: 0792108da6f301abddd4b88320a88bcfc4f5750b