refactored pysa_server.py to stage for Pysa's Language Server specific changes (#409)
Summary:
Expected outcome for this PR is to display and highlight errors for `.pysa` files, via the Pysa VSCode Extension.
This is part of the larger issue of creating a VSCode extension for Pysa in the MLH-Fellowship program listed [here](https://github.com/MLH-Fellowship/pyre-check/issues/4) and has been discussed with gbleaney before.

Possible milestones:
- [X] Make template extension code for Pysa by cloning Pyre
- [X] Add new functions necessary for handling model validation errors to `pysa_server.py` and `persistent.py`
- [X] Connect everything and call `get_invalid_taint_models()` in `query.py`
- [X] Make relevant changes on the extension side
- [x] Clean up the server of old Pyre code

Output so far:
![image](https://user-images.githubusercontent.com/17424778/115070432-8edce100-9f12-11eb-9f3b-020e06226afd.png)

Errors are being calculated but are not being published due to `server_state.opened_documents` list being empty as show in the output. The publishing diagnostics function iterates over this list and sends the errors to VSCode. It's empty right now and hence, the errors are never published. (Opening/closing documents is not being detected)

Pull Request resolved: https://github.com/facebook/pyre-check/pull/409

Reviewed By: arthaud

Differential Revision: D27857795

Pulled By: gbleaney

fbshipit-source-id: b9e8b8026aa48ee5d79db2ed9643ea7e70148e04