Infer: improve the type specificity in TypeInferenceData
Summary:
It came up in review of a separate diff that our types did not clearly
express that attributes and globals always have nonmissing annotations.

This diff fixes that by separating optional type annotations from required
ones. It's a little ugly to need a type synonym for `Type.t`, but as far
as I can tell from skimming the docs there's no way to get a derived
`to_yojson` otherwise, and if the choice is between a type synonym or
a bunch of hand-rolled json serialization I think the former is better.

Reviewed By: dkgi

Differential Revision: D28939568

fbshipit-source-id: ffd76acd4a123cd9b6c9fbe297d7046c8c95ec6f