Infer: Strip out full stub path logic
Summary:
As I was working on porting the full stub path logic (where we extract
existing annotations from code and combine those with infer stubs)
to `infer_v2`, I gradually became aware that this is broken beyond any
easy repair.

The root of the problem is that libcst - or at least to parts of it that
we are relying on - doesn't know enough about import state us to easily
generate correctly-qualified annotations. So, for example, if someone
imports `ArgumentParser` from `argparse` and uses it in an annotation,
we'll generate a stub with a completely unqualified `ArgumentParser` type
and no import.

The problem doesn't really matter if you then inline the stubs, but
in that case there was no point generating full stubs in the first place!

This is a general symptom that infer isn't handling qualifications
for imports in a very reliable way, which is something we'll need to
invest in. But the problem is much harder to address for full stubs than
for ordinary infer, and the feature is so broken right now that I'm almost
100% confident no one is relying on it - it couldn't possibly be useful
on realistic code because it messes up qualifications so badly.

Reviewed By: shannonzhu

Differential Revision: D28996215

fbshipit-source-id: 0283f788b9d800fa1566abef24c75cd75659cd26