Infer: move attribute tests out of inferenceTest.ml
Summary:
This is the first commit moving inference tests out of inferenceTest.ml,
and in the process decoupling the tests from `AnalysisError.t` in favor
of `TypeInferenceData.{Local,Global}Result`.

Almost all of the tests I migrated into `TypeInferenceLocalTest`, which runs
a local inference on just one define. The only exception is the test that
validates we correctly merge a `NoneType` annotation coming from a class
toplevel with an `int` annotation coming from `__init__`; that has to live
in `TypeInferenceGlobalTest` because it involves cross-define merging.

Reviewed By: dkgi

Differential Revision: D28994574

fbshipit-source-id: c00904d400a0112f24520af85b10fdb29d9ce1e0