Infer: jsonify typeInferenceLocalTest
Summary:
I want to port all of `inferenceTest.ml` over to live in
`typeInferenceLocalTest`, which is a required step for decoupling
infer from `AnalysisError.t`.

But the existing `inferenceTest` includes tests of where we *don't*
infer types, not just where we *do*, which the existing test harness
can't really do well. The easiest way to handle this is to just
verify the entire value of a `LocalResult.t` in each test, and it's
more concise / maintainable to do this via json.

So, this commit does the following:
- add a `to_yojson` function to `LocalResult`
- using that, rewrite all the existing tests to validate the entire
  `LocalResult` struct against an expected json blob
  - To keep boilerplate from swamping the actual tests, I tweaked the json
     handling so that we strip out the define if it's a toplevel define

The tests actually get quite a bit simpler :)

Reviewed By: dkgi

Differential Revision: D28983382

fbshipit-source-id: ba8331c25954ad5625e0404d9a9a4af27313f1a5