Reduce the return access path set on joins
Summary:
We currently limit the number of return access paths (i.e, the output path of propagations) to 5 when widening.
I have recently seen the analysis blowing up in function which would create a lot of different return access paths with sequential statements. The widening does not help since it is only applied at the end of the local analysis and at loop heads.

This diff proposes to apply the reduction on joins, which would solve the problem for consecutive statements.

Reviewed By: fahndrich

Differential Revision: D28984599

fbshipit-source-id: 0abc1d7e73761e027b047a2fad1411d02c63e86d