Infer: inferCommand for v2 [3] remove infer from analyzeCommand.ml
Summary:
This is the final diff of a stack decoupling type inference from
running in `analyzeCommand.ml`, we strip out the couple of lines of
type-inference-specific logic.

The benefit of this change is that, particularly combined with last week's
work rewiring the interprocedural top-level framework to explicitly operate
on just one analysis at a time, it will be much easier to parametrize
behavors of the framework itself (e.g. how modules are loaded, precisely what
kind of updates run between fixpoint iterations, what configuration type is
passed to the analysis) in a lightweight way

Reviewed By: kbansal

Differential Revision: D29046918

fbshipit-source-id: 26b49d4f0c2b76ef7f6441e34afa5f55858ab82e