support typing_extensions.TypeAlias as a special form
Summary:
We need to handle `TypeAlias` as a special form in order to
correctly work with the stubs in typeshed:
https://github.com/python/typeshed/blob/master/stdlib/typing_extensions.pyi#L117

This commit updates our faked-out stubs in `test.ml` to match
typeshed better, which breaks our tests by itself, and then adds the
special form to `unannotatedGlobalEnvironment.ml`

NOTE: I'm fairly certain that `typing.TypeAlias`, which becomes available in
3.10, won't work with this code. But we have other logic (like `Type.is_type_alias`)
that also won't work for `typing.TypeAlias`. So it seems better to only handle
`typing_extensions.TypeAlias`, which will still work in 3.10, for now.
We might want to add support when 3.10 is released (or even now, but in
a separate diff)

Reviewed By: grievejia

Differential Revision: D29081810

fbshipit-source-id: d6b22f2274fef697231188670ac27b96ecdf02da