Infer: add stub-writing logic to infer-v2
Summary:
This commit gives infer-v2 the ability to write stub files to disk.

It adds a `--print-only` flag that preserves the current behavior of
logging stubs to stderr and printing the json from ocaml to stdout.

Otherwise, it writes stubs into the same stub directory used by
the original infer command.

We still don't have the ability to do in-place annotations, I'll add
that shortly - going to add one flag at a time to keep things easy
to write, test, and review.

Reviewed By: shannonzhu

Differential Revision: D29090673

fbshipit-source-id: 44a4a4023e1951f35e52b3ddf798315926e85863