Infer: Add --apply-existing-annotations to infer-v2
Summary:
This commit finishes the core logic for infer-v2 by adding the
ability to merge stub files with code files.

We still need to support more flags, and in particular support
one-shot runs where we extract stubs and then apply them immediately,
but all of the libcst logic we need is now hooked up.

Reviewed By: shannonzhu

Differential Revision: D29090671

fbshipit-source-id: 70022476651bac3ad25916db327194936a1becb3