Infer: re-group infer v1 stub-merging logic
Summary:
It's easier for to skim the code if we group all of the code
for merging stubs with `.py` files together.

This commit renames `AnnotatePathArguments` into `AnnotateTask`
and groups all the functionality for combining a `.pyi` and `.py` file
into this class. The reason the task is needed in the first place is so that
we can farm the work out across a multiprocessing pool by making a
list of tasks and iterating over them.

This should be a no-op in terms of actual functionality, it's just
grouping logic together differently.

Reviewed By: shannonzhu

Differential Revision: D29082923

fbshipit-source-id: e0120cc909815ce1124906176463d7750081d97f