Infer: reorganize code
Summary:
Move `infer_v2`s invocation to be next to `infer` instead
of `analyze` (it was next to `analyze` originally because the
arguments were similar when it was invoking analyzeCommand.ml,
but that is no longer true).

Also fix a lint error about an unused import that slipped into
an earlier cleanup commit.

Reviewed By: shannonzhu

Differential Revision: D29090672

fbshipit-source-id: 759d98f0325e3492da711b3db793129f4b7d4b1a