Infer: port integration test to infer-v2
Summary:
In this commit, working toward full compatiblity of infer-v2 with
the existing infer, we add an integration test of the _call_client
calls and flags, in the process adding more logic to infer-v2 so that
it supports the same features supported by infer (v1).

NOTE: this is focusing only on python client compatibility. Some of
these flags may not be correctly interpreted by the ocaml side yet (I
suspect some of the filtering args are not, because at this point we
don't have the ability to restrict qualifiers for interprocedural infer)

Reviewed By: shannonzhu

Differential Revision: D29104466

fbshipit-source-id: d8e5ae007ea7924299e538c5216470750894468b