Ignore decorators when building the call graph
Summary:
When building the call graph for the interprocedural analysis (i.e, the taint analysis and type inference analysis), we resolve calls based on the type of expressions, which often gives us a `Callable(<name>)[...]`. The problem is when functions or methods are decorated with untyped or partially typed decorators, because we end up with `Any`. Until now, we would just mark the call as unsolved and ignore it, which leads to false negatives in pysa.

This diff proposes to fallback to ignoring decorators, and properly adding an edge to the undecorated function or method.
This is implemented using the `GlobalResolution.resolve_exports` and `GlobalResolution.class_definition`, which give syntactical information a symbol or a class.

Reviewed By: dkgi

Differential Revision: D29082748

fbshipit-source-id: e6e7905133187845d4335f646a521467a0255ff3