Infer: tweak inferCommand.ml, support ignore_infer fully
Summary:
I was debugging a path mismatch error that I thought might be due
to `infer` vs `infer-v2` handling config differently, so I decided
to copy-paste the body of `run_infer_v1` up to the creation of the
`configuration` into `run_infer_v2`.

It turns out that wasn't the issue, but I still think this is better
so putting up a commit.

Along the way it made sense to extract the path parsing as a helper
function, and I removed the fail-if-ignore-infer-set logic because
as of D29146638 (https://github.com/facebook/pyre-check/commit/ea861f41c0eba3c69e3d0139bc7a2849862c28b1) it works and is covered by tests

Reviewed By: dkgi

Differential Revision: D29182498

fbshipit-source-id: 345a2fb891bc05339519ac82fe102404c218ccac