Infer: don't overload `in_place_paths`
Summary:
The existing `Infer` command converts the `--in-place`
and `[modify paths]` arguments into a single `in_place_paths`
optional list which is `None` whenever `--in-place` wasn't set.

I copied this over at first for `infer-v2` because I was aiming
for compatibility, but this leads to confusion in code review
because the existing pattern isn't very clear, plus putting
business logic in the `pyre.py` layer is confusing, since that
makes it harder to map the commandline arguments to the actual
`Infer` logic.

So as of this diff, I pass the `--in-place` flag down directly
to `Infer` as a boolean. At this point `infer-v2` doesn't yet
support the `[modify paths]` argument, so we can defer handling
that until later

Reviewed By: dkgi

Differential Revision: D29213269

fbshipit-source-id: e064228d1064e57ebab58d9081da664b00970635