Infer: simplify _annotate_in_place interface
Summary:
There's a slight performance penalty to rglobbing when we
don't have to, which is why I'm guessing the legacy infer has
to copy-pasta'ed codepaths for annotating in place in one shot
versus using the `--annotate-using-existing-stubs` flag.

But I think it's pretty important for maintainability to guarantee
that
```
pyre infer-v2
# followed by
pyre infer-v2 --annotate-using-existing-stubs --in-place
```
produces the same result as
```
pyre infer --in-place
```
and this is easier if we don't have to deal with diverging
relative-to-current-directory versus relative-to-analysis-root
path logic in the two cases, which we can do if we're willing
to just always rglob the stubs directory.

Note that this code change is a no-op at this point anyway, since
we don't yet support one-shot `--in-place` in `infer-v2`.

Reviewed By: dkgi

Differential Revision: D29211730

fbshipit-source-id: bdd3dedac652f8f55dc14f4b6f1bb12230e91d0d